file-processing-exercises/module-1-file-investigation/exercise-1.1-project-handoff/messy-project/requirements.txt
flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
pytest==7.4.3
redis==5.0.1
celery==5.3.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
